requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flaskapp==1.0.5
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathlib==1.0.1
pycodestyle==2.6.0
pymongo==3.11.2
requests==2.25.0
soupsieve==2.1
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
